# docs/swagger.yaml
basePath: /
definitions:
...
  api.InsertWorkerParams:
    properties:
      address:
        type: string
      company:
        type: string
      name:
        type: string
      phone:
        type: string
      taxIdNumber:
        type: string
    type: object
...
  api.UpdateWorkerParams:
    properties:
      address:
        type: string
      company:
        type: string
      name:
        type: string
      phone:
        type: string
      taxIdNumber:
        type: string
    type: object
...
  fiber.Map:
    additionalProperties: true
    type: object
...
  types.Worker:
    properties:
      address:
        type: string
      company:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      taxIdNumber:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
...
  /worker:
    get:
      description: Retrieves a list of workers based on query parameters.
      parameters:
      - description: Worker ID
        in: query
        name: id
        type: string
      - description: Company name
        in: query
        name: company
        type: string
      - description: Worker name
        in: query
        name: name
        type: string
      - description: Phone number
        in: query
        name: phone
        type: string
      - description: Address
        in: query
        name: address
        type: string
      - description: Tax ID number
        in: query
        name: taxIdNumber
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.Worker'
            type: array
      summary: Get workers
      tags:
      - Worker
    post:
      consumes:
      - application/json
      description: Inserts a new worker.
      parameters:
      - description: Worker information
        in: body
        name: worker
        required: true
        schema:
          $ref: '#/definitions/api.InsertWorkerParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Worker'
      summary: Insert worker
      tags:
      - Worker
  /worker/{id}:
    delete:
      description: Deletes a worker by ID.
      parameters:
      - description: Worker ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Delete worker
      tags:
      - Worker
    patch:
      consumes:
      - application/json
      description: Update an existing worker in the system.
      parameters:
      - description: Worker ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated worker details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateWorkerParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Update worker
      tags:
      - Worker
swagger: "2.0"